moved utils to separate utils module
